Handle GITHUB_PATH

+++ install-go.pl
@@ -7,15 +7,24 @@ use 5.010;
 use JSON::PP;
 use HTTP::Tiny;
 use File::Spec;
+use Getopt::Long;
 
-@ARGV == 1 or @ARGV == 2 or die <<EOU;
-usage:
+my $NO_GITHUB_PATH;
+GetOptions('p|dont-alter-github-path' => \$NO_GITHUB_PATH)
+    and (@ARGV == 1 or @ARGV == 2)
+    or die <<EOU;
+usage: $0 [-p|--dont-alter-github-path] GO_VERSION [INSTALL_DIR]
   $0 1.14   [installation_directory/]
   $0 1.9.2  [installation_directory/]
   $0 1.15.x [installation_directory/]
   $0 tip    [installation_directory/]
 
-installation_directory/ defaults to .
+INSTALL_DIR defaults to .
+
+By default, if GITHUB_PATH environment variable exists *AND*
+references a writable file, INSTALL_DIR/go/bin is automatically
+appended to this file.
+-p or --dont-alter-github-path option disables this behavior.
 EOU
 
 
@@ -62,6 +71,7 @@ if ($TARGET eq 'tip')
     if ($goroot)
     {
         install_tip("$goroot/bin/go", $DESTDIR);
+        export_path("$DESTDIR/go/bin");
         exit 0;
     }
 
@@ -77,6 +87,8 @@ my $HTTP = HTTP::Tiny::->new;
 
 link_go_if_available($TARGET, $last_minor, $DESTDIR)
     or install_go(get_url($TARGET, $last_minor), $DESTDIR, $TIP);
+
+export_path("$DESTDIR/go/bin");
 
 exit 0;
 
@@ -144,11 +156,14 @@ sub link_go_if_available
     my $vreg = qr,go[\\/]\Q$full\E[\\/]x64\z,;
     while (my($var, $value) = each %ENV)
     {
-        if ($var =~ /^GOROOT(?:_\d+_\d+_X64)?\z/ and $value =~ $vreg)
+        if ($var =~ /^GOROOT(?:_\d+_\d+_X64)?\z/
+            and $value =~ $vreg
+            and -f -x "$value/bin/go")
         {
             say "Find already installed go version $full";
             rmdir "$dest_dir/go";
             symlink($value, "$dest_dir/go") or die "symlink($value, $dest_dir/go): $!\n";
+            say "go version $full symlinked and available as $dest_dir/go/bin/go";
             return 1;
         }
     }
@@ -168,7 +183,7 @@ sub get_url
 
         say "Check https://golang.org/dl/go$full.$OS-$ARCH.$EXT";
         my $r = $HTTP->head("https://golang.org/dl/go$full.$OS-$ARCH.$EXT");
-        return $r->{url} if $r->{success};
+        return ($r->{url}, $full) if $r->{success};
         say "=> $r->{status}";
 
         if ($r->{status} == 404)
@@ -188,7 +203,7 @@ sub get_url
 
 sub install_go
 {
-    my($url, $dest_dir, $tip) = @_;
+    my($url, $version, $dest_dir, $tip) = @_;
 
     chdir $dest_dir or die "Cannot chdir to $dest_dir: $!\n";
 
@@ -203,7 +218,14 @@ sub install_go
         exe("curl -s \Q$url\E | tar zxf - go/bin go/pkg go/src");
     }
 
-    install_tip("$dest_dir/go/bin/go", $dest_dir) if $tip;
+    if ($tip)
+    {
+        install_tip("$dest_dir/go/bin/go", $dest_dir);
+    }
+    else
+    {
+        say "go $version installed as $dest_dir/go/bin/go";
+    }
 }
 
 sub install_tip
@@ -224,7 +246,6 @@ sub install_tip
     my $final_go = "$dest_dir/go/bin/go";
     if (-e $final_go)
     {
-        say "rename($final_go, $final_go.orig)";
         rename $final_go, "$final_go.orig"
             or die "rename($final_go, $final_go.orig): $!\n";
     }
@@ -233,9 +254,20 @@ sub install_tip
         mkdir_p("$dest_dir/go/bin");
     }
 
-    say "symlink($gotip, $final_go)";
     symlink($gotip, $final_go) or die "symlink($gotip, $final_go): $!\n";
-    #rename $gotip, $final_go or die "rename($gotip, $final_go): $!\n";
+
+    say "go tip installed as $final_go";
+}
+
+sub export_path
+{
+    if (not $NO_GITHUB_PATH
+        and $ENV{GITHUB_PATH}
+        and open(my $fh, '>>', $ENV{GITHUB_PATH}))
+    {
+        say $fh shift;
+        close $fh;
+    }
 }
 
 sub exe
